Split geometry functions into separate file and header
